I wrote about this file.
Recycle Bin: If you deleted a file by mistake, it may still be in the Recycle Bin. You can double-click the Recycle Bin icon on the desktop to open it. If you find the file you need, click and drag it back to the desktop or to a different folder.